7.11—— 1.打算集成ehcache到service方法上 2.今天做了家电service的缓存注解配置（尚未测试） 3.定时清除计划任务更改（只有当重复规则为“OFF”时才清除） 4.所有action只能调用service，不能调用dao，已全部更改

+++ ihakul_server/src/com/xiaoai/TaskJob/XiaoiTaskTimer.java
@@ -33,7 +33,7 @@ public class XiaoiTaskTimer {
 		//判断
 		long currentTimeMillis = System.currentTimeMillis();
 		for (Xiaoitask xiaoitask : list) {
-			if(currentTimeMillis>xiaoitask.getTriggerTime()){
+			if(currentTimeMillis>xiaoitask.getTriggerTime()&&xiaoitask.getRules().equals("OFF")){
 				//清除
 				xiaoitaskService.deleteXiaoitask(xiaoitask);
 				logger.info("计划任务定时清除器被触发清除了:"+xiaoitask.getId());
